+++ internal/domain/entity/transaction.go
@@ -1,8 +1,10 @@
 package entity
 
+// Transaction is entity to represent transaction in payment processing system
+// Please, do not use float32 for money operations in production!
 type Transaction struct {
 	SourceID      *string `json:"source_id"`
 	DestinationID *string `json:"destination_id"`
-	Amount        int64   `json:"amount"`
+	Amount        float32 `json:"amount"`
 	Type          string  `json:"type"`
 }
